+++ Model/Import/Importorder.php
@@ -1422,6 +1422,8 @@ class Importorder extends AbstractModel
      */
     private function createQuote(MagentoCustomer $customer, array $products): Quote
     {
+        $this->backendSession->setBundleItems([]);
+        $this->backendSession->setHasBundleItems(false);
         $customerRepo = $this->customerRepository->getById($customer->getId());
         $currentStore = $this->storeManager->getStore($this->storeId);
         $quote = $this->lengowQuoteFactory->create()
@@ -1500,6 +1502,7 @@ class Importorder extends AbstractModel
             'marketplace' => $this->orderData->marketplace . $paymentInfo,
         ]);
         $quote->collectTotals()->save();
+
         // stop order creation when a quote is empty
         if (!$quote->getAllVisibleItems()) {
             $quote->setIsActive(false);