+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+
+### Bug Fixes
+## [1.7.1](https://github.com/lengow/plugin-magento2/compare/v1.7.0...v1.7.1) (2025-04-16)
+
+
+### Bug Fixes
+
+* **Action:** [PCMT-347] add new field for shipping method for send action in front database and send value in action for mktp ([#57](https://github.com/lengow/plugin-magento2/issues/57)) ([53ee71d](https://github.com/lengow/plugin-magento2/commit/53ee71d46efdc64d9b2fe5c2dc8c856d992ee665))
+* **totals:** [PCMT-287] Fixing the issue with VAT allocation on shipping costs ([#55](https://github.com/lengow/plugin-magento2/issues/55)) ([71098fb](https://github.com/lengow/plugin-magento2/commit/71098fb54c33956e8fcde3437442feb9ac0c31dc))
+
+## [1.7.0](https://github.com/lengow/plugin-magento2/compare/v1.6.10...v1.7.0) (2025-04-01)
+
+
+### Features
+
+* **orders:** [ECP-88] Bundle product manage ([#24](https://github.com/lengow/plugin-magento2/issues/24)) ([5b6f817](https://github.com/lengow/plugin-magento2/commit/5b6f817b15d60df4c37598302aefed9391ab4913))
+
+## [1.6.10](https://github.com/lengow/plugin-magento2/compare/v1.6.9...v1.6.10) (2025-03-20)
+[PCMT-248] : fix config number and dot
+
+* ([276ac5b](https://github.com/lengow/plugin-magento2/commit/276ac5bf62f33ca7170c09a9211bca5183844dbd))
+## [1.6.9](https://github.com/lengow/plugin-magento2/compare/v1.6.8...v1.6.9) (2025-03-20)
+* **import:** [PCMT-191] option for product matching ([#41](https://github.com/lengow/plugin-magento2/issues/41)) ([22d8d94](https://github.com/lengow/plugin-magento2/commit/22d8d9478cd6594206ddc28e585ed86a3c6eb616))
+
 ## [1.6.8](https://github.com/lengow/plugin-magento2/compare/v1.6.7...v1.6.8) (2025-02-24)